yarn run v1.22.19 $ react-scripts build Creating an optimized production build... Compiled successfully.
File sizes after gzip:

  46.62 kB  build\static\js\main.571ee9f9.js
  1.79 kB   build\static\js\787.022287d2.chunk.js
  541 B     build\static\css\main.073c9b0a.css

The project was built assuming it is hosted at /.
You can control this with the homepage field in your package.json.

The build folder is ready to be deployed.
You may serve it with a static server:

  npm install -g serve
  serve -s build

Find out more about deployment here:

  https://cra.link/deployment

Done in 7.79s.でApache上でReactのプレビューを確認